Search: Correctly show smart albums with a "person" filter
Searching for people can be done with two different terms - "person:" and
"subject:", the former being just an alias. However when creating the search
form, the alias was handled before the filter was deserialized, which means that
the "subject" query was not populated.
Switching the order fixes the issuem however it has a side effect - search
terms, which are set directly in the form will override search parameters from
the generic filter, while previously the filter took precedence.