People: Fix setting a person thumbnail
I guess at some point something must have changed, as it was not possible
selecting a new thumbnail for a person/subject. Additionally the person/subject
lookup had to be changed from using query (q) to using the name/slug. This was
needed, as the q-based query can return the wrong subject. Due to the q-query
using LIKE terms, it can match subjects, which are named similar to the currently
open subject, for example Pete and Peter. In order to avoid that, the subject
lookup now uses a name/slug based query.