People: Handle face regions in rotated images
If an image is rotated (meaning it does not have a "normal" orientation), all of
the face regions' information (such as x, y position and width, heigh) are
rotated as well.
PhotoPrism however expects that all face markers are in "normal", non-rotated
coordinates, so we need to convert the face regions if the image is rotated.