People: Import exif face regions for new files
The exif face regions import only worked for already indexed files, due to
erroneous assumption. For new files the file uid is determined at the very end
of the import process, when the file is saved. This means that no new markers
could be created from the exif face regions - instead a cryptic message was
logged. Deferring the marker creation to the end of the import process fixes the
issue.