People: Detect low quality faces based on thumbnails
Use the thumbnails to detect the low quality faces instead of using the whole
image. This is consistent with all the other places where faces are detected.

This should fix the low quality face detection for rotated images, as the
previous approach did not auto-rotate them, so the face detection was often
garbage for rotated images.