People: Remove support for comparing faces of different scales
In the first implementation draft, we had face regions of different scales -
some were created from the original image and some - from the thumbnail. As
such, the regions had different scales, which had to be taken into account when
comparing them.
Now all of the regions are of the same scale - that of the original image, so
the normalization process is superfluous.