People: Fallback scenario for angled face regions
The most probable reason why a manually labelled face region could not be
detected by PhotoPrism seems to be that the face region is at a weird angle. In
order to support such face regions, a fallback has been added, which kicks in
when no embeddings can be computed for a manually labelled face region. In this
case, the fallback runs the face detection at a couple of hardcoded angles
(close to 90°, 180° and 270°) and checks whether a face has been detected, which
significatly overlaps with the manual face region.